emuNAND RAW1 support
Just a quick band aid fix for people who has emuNAND and found out it doesn't work and their "Nintendo" folder on "/emuMMC/RAW1/Nintendo/". If you have other emuNANDs on your sd card, you can just edit the address inside .py file until a proper gui config application is created if there will be one. Just run this file in PyNX with the E1 at the end.

ps. idk if im spot on the terms but yeah as i was away from the scene for a long time.